CLA Enforcer
It's important that we make contributors sign the Contributors
License Agreement, but unfortunately none of the available services
such as CLA Assistant or CLAHub fit our needs.

We don't need fancy feature at the moment just a webhook that receives
pull request events and creates a status check that is passed when
the contributors signs the agreement.

Of course we also implement a web frontend where the contributor
can actually read and sign the agreement, and an administrative
interface where we can download the list of contributors that agreed.